# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: Hotel Booking API specification
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /hotels:
    get:
      operationId: listHotels
      tags: [hotels]
      summary: List/search hotels
      parameters:
        - name: search
          in: query
          schema:
            type: string
        - name: city
          in: query
          schema:
            type: string
        - name: minPrice
          in: query
          schema:
            type: number
        - name: maxPrice
          in: query
          schema:
            type: number
        - name: minRating
          in: query
          schema:
            type: number
        - name: stars
          in: query
          schema:
            type: integer
        - name: amenities
          in: query
          schema:
            type: string
        - name: sortBy
          in: query
          schema:
            type: string
            enum: [price, rating, popularity]
        - name: sortOrder
          in: query
          schema:
            type: string
            enum: [asc, desc]
      responses:
        "200":
          description: List of hotels
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Hotel"

    post:
      operationId: createHotel
      tags: [hotels]
      summary: Create a new hotel (admin only)
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateHotelBody"
      responses:
        "201":
          description: Hotel created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Hotel"
        "403":
          description: Forbidden
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
...
  /hotels/{id}:
    get:
      operationId: getHotel
      tags: [hotels]
      summary: Get a hotel by ID
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Hotel details
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/HotelWithRooms"
        "404":
          description: Hotel not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"

    patch:
      operationId: updateHotel
      tags: [hotels]
      summary: Update a hotel (admin only)
      security:
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateHotelBody"
      responses:
        "200":
          description: Updated hotel
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Hotel"
        "404":
          description: Hotel not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"

    delete:
      operationId: deleteHotel
      tags: [hotels]
      summary: Delete a hotel (admin only)
      security:
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Hotel deleted
        "404":
          description: Hotel not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /hotels/{id}/similar:
    get:
      operationId: getSimilarHotels
      tags: [hotels]
      summary: Get similar hotels
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Similar hotels
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Hotel"

  # Rooms
...
  /hotels/{hotelId}/reviews:
    get:
      operationId: getHotelReviews
      tags: [reviews]
      summary: Get reviews for a hotel
      parameters:
        - name: hotelId
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: List of reviews
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/ReviewWithUser"

    post:
      operationId: createReview
      tags: [reviews]
      summary: Create a review for a hotel
      security:
        - bearerAuth: []
      parameters:
        - name: hotelId
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateReviewBody"
      responses:
        "201":
          description: Review created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ReviewWithUser"
        "400":
          description: You must have a completed booking to review
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /reviews/{id}:
    delete:
      operationId: deleteReview
      tags: [reviews]
      summary: Delete a review (admin or owner)
      security:
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Review deleted

  # Favorites
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
...
    ErrorResponse:
      type: object
      properties:
        error:
          type: string
      required:
        - error
...
    User:
      type: object
      properties:
        id:
          type: integer
        email:
          type: string
        name:
          type: string
        role:
          type: string
          enum: [USER, ADMIN]
        createdAt:
          type: string
          format: date-time
      required:
        - id
        - email
        - name
        - role
        - createdAt
...
    Hotel:
      type: object
      properties:
        id:
          type: integer
        name:
          type: string
        description:
          type: string
        city:
          type: string
        address:
          type: string
        rating:
          type: number
        stars:
          type: integer
        amenities:
          type: array
          items:
            type: string
        images:
          type: array
          items:
            type: string
        reviewCount:
          type: integer
        minPrice:
          type: ["number", "null"]
        createdAt:
          type: string
          format: date-time
      required:
        - id
        - name
        - description
        - city
        - address
        - rating
        - stars
        - amenities
        - images
        - reviewCount
        - createdAt
    HotelWithRooms:
      type: object
      properties:
        id:
          type: integer
        name:
          type: string
        description:
          type: string
        city:
          type: string
        address:
          type: string
        rating:
          type: number
        stars:
          type: integer
        amenities:
          type: array
          items:
            type: string
        images:
          type: array
          items:
            type: string
        reviewCount:
          type: integer
        minPrice:
          type: ["number", "null"]
        createdAt:
          type: string
          format: date-time
        rooms:
          type: array
          items:
            $ref: "#/components/schemas/Room"
      required:
        - id
        - name
        - description
        - city
        - address
        - rating
        - stars
        - amenities
        - images
        - reviewCount
        - createdAt
        - rooms
    CreateHotelBody:
      type: object
      properties:
        name:
          type: string
        description:
          type: string
        city:
          type: string
        address:
          type: string
        stars:
          type: integer
        amenities:
          type: array
          items:
            type: string
        images:
          type: array
          items:
            type: string
      required:
        - name
        - description
        - city
        - address
        - stars
    UpdateHotelBody:
      type: object
      properties:
        name:
          type: string
        description:
          type: string
        city:
          type: string
        address:
          type: string
        stars:
          type: integer
        amenities:
          type: array
          items:
            type: string
        images:
          type: array
          items:
            type: string
...
    Room:
      type: object
      properties:
        id:
          type: integer
        hotelId:
          type: integer
        type:
          type: string
          enum: [single, double, deluxe, suite]
        price:
          type: number
        guests:
          type: integer
        description:
          type: string
        images:
          type: array
          items:
            type: string
        createdAt:
          type: string
          format: date-time
      required:
        - id
        - hotelId
        - type
        - price
        - guests
        - description
        - images
        - createdAt
...
    ReviewWithUser:
      type: object
      properties:
        id:
          type: integer
        userId:
          type: integer
        hotelId:
          type: integer
        rating:
          type: integer
          minimum: 1
          maximum: 5
        comment:
          type: string
        createdAt:
          type: string
          format: date-time
        user:
          $ref: "#/components/schemas/User"
      required:
        - id
        - userId
        - hotelId
        - rating
        - comment
        - createdAt
        - user
    CreateReviewBody:
      type: object
      properties:
        rating:
          type: integer
          minimum: 1
          maximum: 5
        comment:
          type: string
      required:
        - rating
        - comment
...